OPTIMIZED-VPS-FLASK/requirements.txt
flask>=3.0.0
gunicorn>=22.0.0
openai>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
requests-oauthlib>=1.3.1
